+++ internal/base/mod.go
@@ -1,11 +1,5 @@
 package base
 
-import (
-	"strconv"
-	"strings"
-	"time"
-)
-
 // Mod represents a collection of commands and passives.
 type Mod interface {
 	Name() string
@@ -45,44 +39,4 @@ type ModPassive struct {
 	AllowedTypes MessageType
 	Enabled      bool
 	Run          func(*DiscordMessage) `json:"-"`
-}
-
-// Min returns the minimum of two numbers. Convenience function.
-func Min(a, b int) int {
-	if a < b {
-		return a
-	}
-	return b
-}
-
-// Max returns the maximum of two numbers. Convenience function.
-func Max(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
-}
-
-// Clamp clamps a number between a lower and upper limit. Convenience function.
-func Clamp(lower, upper, n int) int {
-	return Max(lower, Min(upper, n))
-}
-
-// IDToTimestamp converts a discord ID to a timestamp
-func IDToTimestamp(idStr string) time.Time {
-	id, err := strconv.ParseInt(idStr, 10, 64)
-	if err != nil {
-		return time.Now()
-	}
-
-	id = ((id >> 22) + 1420070400000) / 1000
-	return time.Unix(id, 0)
-}
-
-func TrimUserID(id string) string {
-
-	id = strings.TrimPrefix(id, "<@!")
-	id = strings.TrimSuffix(id, ">")
-
-	return id
 }
